+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.9.1.0
+
+- added available/in logic skulltula counters
+- added more logic tricks
+
 ## 2.9.0.0
 
 - replaced ice traps with triforce pieces